requirements.txt
python-nginx
cloudflare
PyNamecheap